Добавит разбиение JS на модули
